requirements.txt
Django==4.0.4
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-filter==21.1
Markdown==3.3.6
drf-yasg==1.20.0